refactor: table of shifts now built in separate component
