+++ web/src/domains/shifts/components/ShiftTable.vue
@@ -1,42 +1,38 @@
 <template>
     <loading-container :loading="false">
-        <div>
-            <h1>{{ team.name }}</h1>
-            <p>{{ team.description }}</p>
-        </div>
-        <div>
-            <h1>Shifts</h1>
-            <v-row>
-                <v-table>
-                <thead>
-                    <tr class="text-left">
-                        <th>Name</th>
-                        <th>Day</th>
-                        <th>Time</th>
-                        <th>Length</th>
-                        <th>Signups</th>
-                        <th>Action</th>
-                    </tr>
-                </thead>
-                    <tbody>
-                        <shift-table-row v-for="shift in shifts" :key="shift"
-                            @shift-action="shiftAction"
-                            :name="shift.name"
-                            :start="shift.start"
-                            :end="shift.end"
-                            :duration="shift.duration"
-                            :description="shift.description"
-                            :signups="shift.signups ?? 0"
-                            :capacity="shift.capacity"
-                            :id="shift.id"
-                            :day="shift.day"
-                            :button="shift.button"
-                            />
-                    </tbody>
-            </v-table>
-            </v-row>
-        </div>
-    </loading-container>
+    <div>
+        <h1>Shifts</h1>
+        <v-row>
+            <v-table>
+            <thead>
+                <tr class="text-left">
+                    <th>Name</th>
+                    <th>Day</th>
+                    <th>Time</th>
+                    <th>Length</th>
+                    <th>Signups</th>
+                    <th>Action</th>
+                </tr>
+            </thead>
+                <tbody>
+                    <shift-table-row v-for="shift in shifts" :key="shift"
+                        @shift-action="shiftAction"
+                        :name="shift.name"
+                        :start="shift.start"
+                        :end="shift.end"
+                        :duration="shift.duration"
+                        :description="shift.description"
+                        :signups="shift.signups ?? 0"
+                        :capacity="shift.capacity"
+                        :id="shift.id"
+                        :day="shift.day"
+                        :button="buildButton(shift)"
+                        />
+                </tbody>
+        </v-table>
+        </v-row>
+    </div>
+</loading-container>
 </template>
 <script>
 import { client } from '../../../../api-client/client'
@@ -46,10 +42,14 @@ import LoadingContainer from '@/domains/shared/LoadingContainer.vue'
 
 export default {
     props: {
-        teamId: {
-            type: String,
+        shifts: {
+            type: Array,
             required: true
-        }
+        },
+        userShifts: {
+            type: Array,
+            required: true
+        },
     },
     components: {
         ShiftTableRow,
@@ -58,30 +58,11 @@ export default {
         ,
     data() {
         return {
-            team: {},
-            shifts: [],
-            userShifts: [],
             userStore: initUserStore(),
             isLoading: true
         }
     },
     methods: {
-        async getTeamById(teamId){
-            const team = await client.teams.getTeamById(teamId)
-            this.team = team.data
-        },
-        async getTeamShifts(teamId){
-            let shifts = await client.teams.getShifts(teamId)
-            shifts = shifts.data
-            shifts.map(shift => {
-                shift.button = this.buildButton(shift)
-            })
-            this.shifts = shifts
-        },
-        async getUserShifts(){
-            const shifts = await client.users.getShifts(this.userStore.userId)
-            this.userShifts = shifts.data
-        },
         buildButton(shift){
             if (this.userShifts.some(userShift => userShift.id === shift.id)) {
                 return {
@@ -116,7 +97,7 @@ export default {
                 default:
                     break
             }
-            await this.load()
+            this.$emit('shift-action')
         },
         async signup(shiftId){
             this.isLoading = true
@@ -135,20 +116,7 @@ export default {
             this.isLoading = false
 
         },
-        async load() {
-
-            await Promise.all([
-                await this.getTeamById(this.teamId),
-                await this.getUserShifts(),
-                await this.getTeamShifts(this.teamId),
-            ])
-        },
     },
-    async created() {
-        await this.load()
-
-
-    }
 }
 </script>
 <style lang="scss" scoped>
